Renamed AndroidAudio and AndroidInput implementations to follow naming convention
