+++ backends/gdx-backend-android/src/com/badlogic/gdx/backends/android/AndroidApplication.java
@@ -496,11 +496,11 @@ public class AndroidApplication extends Activity implements AndroidApplicationBa
 
 	@Override
 	public AndroidAudio createAudio (Context context, AndroidApplicationConfiguration config) {
-		return new AndroidAudioImpl(context, config);
+		return new DefaultAndroidAudio(context, config);
 	}
 
 	@Override
 	public AndroidInput createInput (Application activity, Context context, Object view, AndroidApplicationConfiguration config) {
-		return new AndroidInputImpl(this, this, graphics.view, config);
+		return new DefaultAndroidInput(this, this, graphics.view, config);
 	}
 }
